pyproject.toml
[project]
name = "cc-plugin-to-codex"
version = "0.1.0"
description = "Bridge Claude Code marketplace plugins into Codex (skills + subagents)"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "fangzzzjjj", email = "fangzzzjjj@gmail.com"}]
keywords = ["codex", "claude-code", "plugin", "cli", "bridge", "marketplace"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "typer>=0.12,<1.0",
    "questionary>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",
    "tomli-w>=1.0,<2.0",
    "rich>=13.0,<15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<10.0",
    "pytest-cov>=5.0,<8.0",
]

[project.urls]
Homepage = "https://github.com/fangzzzjjj/cc-plugin-to-codex"
Repository = "https://github.com/fangzzzjjj/cc-plugin-to-codex"
Issues = "https://github.com/fangzzzjjj/cc-plugin-to-codex/issues"
Changelog = "https://github.com/fangzzzjjj/cc-plugin-to-codex/blob/main/CHANGELOG.md"

[project.scripts]
cc2codex = "cc_plugin_to_codex.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cc_plugin_to_codex"]

[tool.hatch.build.targets.sdist]
include = ["src/cc_plugin_to_codex", "README.md", "LICENSE", "CHANGELOG.md", "pyproject.toml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
